feat: rename the file index.d.ts to types.d.ts
